requirements.txt
numpy==1.23.3
pandas==1.5.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
typing_extensions==4.3.0
python-dotenv~=0.21.0